+++ packages/adp-tooling/src/base/prompt.ts
@@ -6,7 +6,6 @@ import type { Logger } from '@sap-ux/logger';
 import type { UI5FlexLayer } from '@sap-ux/project-access';
 import type { AppIndex } from '@sap-ux/axios-extension';
 import { v4 as uuidv4 } from 'uuid';
-
 import { getPackageJSONInfo } from '../writer/project-utils';
 import { isNotEmptyString, isValidSapClient } from './validators';