+++ packages/adp-tooling/src/prompts/index.ts
@@ -1,3 +1,4 @@
 export { getPrompts as getPromptsForChangeDataSource } from './change-data-source';
 export { default as ProjectPrompter } from './creation';
+export { getPrompts as getPromptsForAddAnnotationsToOData } from './add-annotations-to-odata';
 export * from './creation/prompt-helpers';