+++ packages/axios-extension/src/abap/types/adt-types.ts
@@ -145,5 +145,11 @@ export type BusinessObject = {
     description?: string;
 };
 
+export type AbapCDSView = {
+    name: string;
+    uri: string;
+    description?: string;
+};
+
 export type PublishResponse = { SEVERITY: string; SHORT_TEXT: string; LONG_TEXT: string };
 export type ValidationResponse = { severity: string; short_text: string; long_text: string };