+++ packages/inquirer-common/src/types.ts
@@ -60,7 +60,10 @@ export interface GuiOptions {
     type?: string;
 }
 
-export type PromptSeverityMessage = (input?: unknown, previousAnswers?: Answers) => IMessageSeverity | undefined;
+export type PromptSeverityMessage = (
+    input?: unknown,
+    previousAnswers?: Answers
+) => IMessageSeverity | undefined | Promise<IMessageSeverity | undefined>;
 
 export type YUIQuestion<A extends Answers = Answers> = Question<A> & {
     name: string;
@@ -77,6 +80,7 @@ export interface FileBrowserQuestion<A extends Answers = Answers> extends BaseIn
 export interface ListQuestion<A extends Answers = Answers> extends BaseListQuestion<A> {
     name: YUIQuestion['name'];
     guiOptions?: YUIQuestion['guiOptions'];
+    additionalMessages?: YUIQuestion['additionalMessages'];
 }
 
 export interface ConfirmQuestion<A extends Answers = Answers> extends BaseConfirmQuestion<A> {