+++ packages/annotation-generator/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/annotation-generator
 
+## 0.1.12
+
+### Patch Changes
+
+-   Updated dependencies [a9fac04]
+    -   @sap-ux/project-access@1.25.6
+    -   @sap-ux/fiori-annotation-api@0.1.22
+
+## 0.1.11
+
+### Patch Changes
+
+-   Updated dependencies [421f3ca]
+    -   @sap-ux/project-access@1.25.5
+    -   @sap-ux/fiori-annotation-api@0.1.21
+
 ## 0.1.10
 
 ### Patch Changes