+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.5.9
+
+### Patch Changes
+
+-   Updated dependencies [a9fac04]
+    -   @sap-ux/project-access@1.25.6
+
+## 0.5.8
+
+### Patch Changes
+
+-   Updated dependencies [108336f]
+    -   @sap-ux/project-input-validator@0.3.1
+    -   @sap-ux/inquirer-common@0.4.1
+
+## 0.5.7
+
+### Patch Changes
+
+-   Updated dependencies [421f3ca]
+    -   @sap-ux/project-access@1.25.5
+
 ## 0.5.6
 
 ### Patch Changes